pyproject.toml
[tool.poetry]
name = "tgchan-bookmark-bot"
version = "0.1.0"
description = "After you save a page as bookmark, post it in the Telegram channel"
authors = ["myl7 <myl@myl.moe>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.27.1"
python-telegram-bot = "^13.11"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
